Implement CPU utilization
CPU utilization is recoded each time a VM serves a request. The start
and stop time of request service are saved in a deque so they can be
later used for the evaluation of the CPU utilization, as the average CPU
utilization in the last (float) x seconds.  In a VPM each VM has the
same amount of allocated resources, then more VMs means more computation
capacity, then the CPU utilization depends on the number of VMs.

If we consider an interval in which we want to evaluate a CPU
utilization, in this interval VM number can change. Because of this, we
must
- divide the interval in subintervals in which the number of VM is
	constant
- evaluate the CPU utilization in each subinterval
	(see __usage_interval_constant_vm())
- evaluate the time weighted average of CPU utilization over the
	complete interval: this is our CPU utilization
	(see __usage_interval())

Each time a change in number of VMs occurs, it is recorded in a deque
because we need them to split the intervals in subintervals, as
mentioned before.