Introduce WebRequest Event subclass
The previous implementation of the service of a request was modelled as
a process. A client had to provide the server with the parameters of the
request and the server returned the process that the client had to
yield.

That implementation was difficult to scale: since we are going to
introduce load balancers, there will be more layers of components that
will route the request.

This is easier to handle if requests are modelled as Event objects.
Clients will provide either the server or the load balancer with a
reference of the WebRequest event. Then they yield it.
On the other hand the server or the load balancer will handle the request
conveniently and either call `succeed()` or `fail()` on the WebRequest.