Create abstract load balancer
Load balancers can be either local or global. Regardless of the type,
they share the admission mechanism, which is a loop that routes the
requests to the appropriate server, and the request submission.

The variable part is how the requests are routed, hence its
implementation is left to the child class.