Implement the Local Load Balancer
I introduced the LocalLoadBalancer, that inherits from LoadBalacer and
implements the add_server() and the route() functions.

The route() function does not require to make a decision on the server
to which to forward the request, because it is only one.

I renamed the Server.request member function to Server.submit_request so
that you can x.submit_request( <a request> ) regardless of x beeing a
Server or a LoadBalancer.