Implement Global Load Balancer
The global load balancer can forward requests to multiple servers (or
multiple load balancers in the future). To do so, for every request
there must be a way to choose a server which to forward the request to.

I implemented a "turning" policy and a "least CPU loaded" policy. The
one to choose is up to the user. It can be configured at initialization
time or later through the `route__config()` function.