Make Server a subclass of UsageManager
In order to get the usage of a server, it was necessary to access the
internal Server reference to its usage manager. But usage requests
should be part of the Server interface and should not require access to
its members.

Now the Server class is a subclass of UsageManager and the usage request
methods are now directly part of the Server interface.