Fix server capacity shrink issue
If we are shrinking the capacity of the server, but the VMs that are busy
are less than the capacity, then the capacity change takes immediately
effect. Contrarily to what was previously logged into the capacity
changes.

If the new capacity is greater than the VMs that are busy, then it is the
new effective capacity. If the new capacity is less than the busy VMs
then the effective capacity is the number of busy VMs. Once the requests
that are served by the VMs are completed, the serving process itself
will log the effective capacity if the busy VM count is greater than the
desired capacity before releasing the VM