Add recursive least loaded routing
To check which is the least loaded portion of the cloud, we ask the
lower level load balancers to evaluate the average usage of the servers
in the subtree.

Note that in the interval in which the usage is evaluated, the number of
virtual machines in the server could not be constant. To weight the
usage, by the virtual machine numbers it is therefore an approximation
of the average usage.