Switch to environment based time monitoring
Capacity changes and resource acquisition and release were recorded
using the time module. This solution was not appropriate as the
environment time and the clock time may not coincide. While all events
are scheduled at a certain simulation time, the simulation can be
stopped, resumed, or even fastened for Realtime Environments, so the
time the events are scheduled is not guaranteed to be synchronous with
the system time.