Fix UsageManager bugs and add new example
In the context manager of UsageManager we should set a tuple when
logging the utilization, because it will be later indexed and a float
cannot be indexed.

We should also check that utilization is not evaluated in a null
interval.

The new example shows how to add a control algorithm to the server farm.
Note that the controller will end when the request generator will
terminate, thanks to the condition on the while loop.